+++ packages/base-editor/src/components/editor/Editor.js
@@ -3,6 +3,7 @@ import { UIElement, useStore } from "@elf-framework/sapa";
 import { EditorContext } from "../../managers/EditorContext";
 
 const KEY_EDITOR = "editor";
+const KEY_EDITOR_OPTION = "editorOption";
 
 export class EditorPlugin {
   constructor(editor, props = {}) {
@@ -19,8 +20,49 @@ export class EditorPlugin {
   deactivate() {}
 }
 
+/**
+ * EditorContext 조회
+ */
 export function useEditor() {
   return useStore(KEY_EDITOR);
+}
+
+/**
+ * Editor Option 조회
+ *
+ */
+export function useEditorOption(key) {
+  return useStore(KEY_EDITOR_OPTION)?.[key];
+}
+
+/**
+ * Editor Config 조회
+ */
+export function useConfig(key) {
+  return useEditor()?.configs?.get(key);
+}
+
+/**
+ * Editor Config 값 설정
+ *
+ */
+export function useSetConfig(key, value) {
+  return useEditor()?.configs?.set(key, value);
+}
+
+/**
+ * Editor Command 실행
+ *
+ */
+export async function useCommand(key, ...args) {
+  return await useEditor()?.commands?.emit(key, ...args);
+}
+
+/**
+ * 다국어 메세지 조회
+ */
+export function useI18n(key, params = {}) {
+  return useEditor()?.i18n?.get(key, params);
 }
 
 /**
@@ -41,6 +83,7 @@ export class Editor extends UIElement {
     this.$editor = new EditorContext(this, this.props);
 
     this.$store.set(KEY_EDITOR, this.$editor);
+    this.$store.set(KEY_EDITOR_OPTION, this.props);
 
     this.activate();
   }