+++ packages/sapa/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @elf-framework/sapa
 
+## 2.0.8
+
+### Patch Changes
+
+- Support useSubscribe, useEmit
+
 ## 2.0.7
 
 ### Patch Changes